+++ CHANGES.md
@@ -4,7 +4,7 @@
 0.9.3 (unreleased)
 ------------------
 
-- Nothing changed yet.
+- Solved aiohttp 'Unclosed client session' warning.
 
 
 0.9.2 (2023-11-23)